Move base classes to common graph dir
