+++ checkov/graph/checks/checks_infra/solvers/attribute_queries/base_attribute_solver.py
