+++ checkov/graph/checks/checks_infra/solvers/connection_queries/base_connection_solver.py
