+++ checkov/graph/terraform/checks/checks_infra/solvers/attribute_queries/not_ending_with_attribute_solver.py
@@ -1,4 +1,4 @@
-from checkov.graph.terraform.checks.checks_infra.solvers.attribute_queries.base_attribute_solver import BaseAttributeSolver
+from checkov.graph.checks.checks_infra.solvers.attribute_queries.base_attribute_solver import BaseAttributeSolver
 
 
 class NotEndingWithAttributeSolver(BaseAttributeSolver):
